# openapi.yaml
---
components:
  schemas:
    getFileResponse:
      properties:
        fileData:
          description: File data.
          type: object
      type: object
    getRepoResponse:
      properties:
        repoData:
          description: Repo data.
          type: object
      type: object
    listFilesResponse:
      properties:
        files:
          description: List of file names.
          items:
            type: string
          type: array
      type: object
info:
  description: A plugin that allows the user to get data from GitHub using ChatGPT. If you do not know the user's username, ask them first before making queries to the plugin. Otherwise, use the username "global".
  title: GitHub Plugin
  version: v1
openapi: 3.0.1
paths:
  /file/{username}/{repo_name}/{file_name}:
    get:
      operationId: getFile
      parameters:
        - description: The name of the repo.
          in: path
          name: repo_name
          required: true
          schema:
            type: string
        - description: The name of the user.
          in: path
          name: username
          required: true
          schema:
            type: string
        - description: The name of the file.
          in: path
          name: file_name
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  fileData:
                    description: File data.
                    type: object
                type: object
          description: OK
      summary: Get file data from GitHub
  /list-files/{owner}/{repo}/{path}:
    get:
      operationId: listFiles
      parameters:
        - description: The name of the repo owner.
          in: path
          name: owner
          required: true
          schema:
            type: string
        - description: The name of the repo.
          in: path
          name: repo
          required: true
          schema:
            type: string
        - default: ''
          description: The path to the file.
          in: path
          name: path
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  files:
                    description: List of file names.
                    items:
                      type: string
                    type: array
                type: object
          description: OK
      summary: List files in a repo
  /repo/{username}/{repo_name}:
    get:
      operationId: getRepo
      parameters:
        - description: The name of the repo.
          in: path
          name: repo_name
          required: true
          schema:
            type: string
        - description: The name of the user.
          in: path
          name: username
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  repoData:
                    description: Repo data.
                    type: object
                type: object
          description: OK
      summary: Get repo data from GitHub
servers:
  - url: http://localhost:5003